requirements.txt
streamlit
python-dotenv
langchain-openai==0.2.6
langchain-core>=0.3.15,<0.4.0